pyproject.toml
[project]
name = "raven-zero"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"  # Cambiado de 3.14 a 3.13 para mayor estabilidad
dependencies = [
    "fastapi[standard]>=0.115.0",
    "redis>=5.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "structlog>=23.2.0",
    "apscheduler>=3.10.0",
    "aiofiles>=23.0.0",
    "python-magic-bin>=0.4.14",
    "cryptography>=46.0.3",
    "logging>=0.4.9.6",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "fakeredis>=2.20.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-redis>=4.5.0",
    "types-python-dateutil>=2.8.19",
    "types-pytz>=2023.3.0",
    "types-setuptools>=75.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N"] # I = Ordenar imports, N = Nombramiento de variables
